Add screen-shot to the README.
I honestly think we're getting somewhere solid and useful. There's still
a long road until I'd consider this a testing framework, but honestly,
just looking at it is like a breath of fresh air in an ecosystem like
D's (look at the difference of this output and dmd's for instance).

[ci skip]