Always use `cwrite..` over `write..`.
This should keep us from having too many windows compatibility problems,
and is mainly due to the fact that, since the latest changes, I've
started using a colorized string as a title (the `rootSuite.title` is
colorized yellow, to differentiate it from the user defined suites).

We can obviously provide a better implementation than using the
hand-rolled `rootSuite` hack, but I'd say this works for our initial
draft.