Start adding new structure for concurrency
This starts adding what I think is the required structure for decoupling
reporting from running and running concurrently. It adds a `core`,
delegation module and an `interfaces.bdd` base interface (to be
implemented). It also starts drafting the reporter module (which may or
may not be factored into several different modules in the future) and
tweaks our tests' data modeling.